+++ src/com/example/SanAngeles/DemoActivity.java
@@ -76,9 +76,7 @@ public class DemoActivity extends Activity {
 
     @Override
     protected void onDestroy() {
-        if (mGLView != null) {
-            mGLView.releaseNativeResources();
-        }
+        mGLView = null;
         super.onDestroy();
     }
 
@@ -112,8 +110,9 @@ class DemoGLSurfaceView extends GLSurfaceView {
 
     @Override
     public void onPause() {
-        super.onPause();
         nativePause();
+        releaseNativeResources();
+        super.onPause();
     }
 
     @Override
@@ -123,7 +122,11 @@ class DemoGLSurfaceView extends GLSurfaceView {
     }
 
     public void releaseNativeResources() {
-        mRenderer.releaseNativeResources();
+        queueEvent(new Runnable() {
+            public void run() {
+                mRenderer.releaseNativeResources();
+            }
+        });
     }